pyproject.toml
[project]
name = "pytest-appium-testing"
version = "0.1.0"
description = ""
package-mode = false
authors = [
    {name = "harden hsieh",email = "harden.cc.hsieh@viewsonic.com"}
]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "appium-python-client (>=5.2.6,<6.0.0)"
]


[dependency-groups]
dev = [
    "black (>=26.1.0,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
(
    .git
  | .venv
)
'''

[tool.isort]
profile = "black"
line_length = 120
skip = [
    '.bzr',
    '.direnv',
    '.eggs',
    '.git',
    '.hg',
    '.mypy_cache',
    '.nox',
    '.pants.d',
    '.svn',
    '.tox',
    '.venv',
    '__pypackages__',
    '_build',
    'buck-out',
    'build',
    'dist',
    'venv',
    '__pycache__',
    'report',
]
